+++ docs/changelog.md
@@ -13,7 +13,7 @@ Here you can find all the changelogs for `hikari-arc`.
 
 - Add [loops](./guides/loops.md). Loops can be used to repeatedly call a given coroutine function with a specific interval or cron set.
 - Add [concurrency limiters](./guides/concurrency_limiting.md). Concurrency limiters can be used to prevent users from invoking a command that already has a specific amount of instances running.
-- Add `arc.StartedEvent` to gateway clients, this fires after command syncing has completed.
+- Add `arc.StartedEvent` and `arc.StoppingEvent` to gateway clients to enable managing lifecycle via events.
 - Fix command groups always being republished when command syncing.
 
 ## v0.6.0